'JWT_auth'

+++ backend/controllers/goalController.js
@@ -2,12 +2,13 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 // @desc Get goals
 // @route Get /api/goals
 // @access Private
 const getGoals = asyncHandler(async(req,res)=>{
-const goals = await Goal.find();
+const goals = await Goal.find({user: req.user.id});
 
 
     res.status(200).json(goals);
@@ -25,7 +26,8 @@ const setGoal = asyncHandler(async(req,res)=>{
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     });
 
     res.status(200).json(goal);
@@ -40,6 +42,20 @@ const updateGoals = asyncHandler(async(req,res)=>{
     if(!goal){
         res.status(400);
         throw new Error('Goal not found');
+    }
+
+    const user = await User.findById(req.user.id);
+
+    // check if user exist
+    if(!user){
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    // Make sure the loggged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401);
+        throw new Error('User not authorized');
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body , {
@@ -58,6 +74,20 @@ const deleteGoals = asyncHandler(async(req,res)=>{
     if(!goal){
         res.status(400);
         throw new Error('Goal not found');
+    }
+
+    const user = await User.findById(req.user.id);
+
+    // check if user exist
+    if(!user){
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    // Make sure the loggged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401);
+        throw new Error('User not authorized');
     }
     
     await goal.deleteOne()
